tools,tests: fix 64-bit Linux -Wformat and -Wcomment warnings
On LP64 (Linux/gcc) uint64_t is `unsigned long`, so the locus-gguf
tensor printf's %llu specifiers mismatched their arguments; cast to
unsigned long long to match %llu on every platform (no-op on macOS
where uint64_t is already unsigned long long). Also drop a stray
line-continuation backslash that turned a // comment into a spliced
multi-line comment. Both were warning-clean on the peer's arm64/clang
build but not under a clean gcc build; needed before enabling CUDA.

+++ tools/gguf_info_main.cpp
@@ -41,8 +41,12 @@ int main(int argc, char** argv) {
             std::printf(
                 "%s ne=[%llu,%llu,%llu,%llu] type=%u "
                 "bytes=%llu first=%02x\n",
-                t->name.c_str(), t->ne[0], t->ne[1], t->ne[2],
-                t->ne[3], (unsigned)t->type, t->nbytes,
+                t->name.c_str(),
+                (unsigned long long)t->ne[0],
+                (unsigned long long)t->ne[1],
+                (unsigned long long)t->ne[2],
+                (unsigned long long)t->ne[3], (unsigned)t->type,
+                (unsigned long long)t->nbytes,
                 (unsigned)g.tensor_data(*t)[0]);
         } else if (cmd == "tokens" && argc > 3) {
             auto tok = locus::tok::tokenizer_from_gguf(g);
